Generalize the implementation for Title
