+++ bokehjs/src/lib/models/annotations/label.ts
@@ -2,23 +2,18 @@ import {TextAnnotation, TextAnnotationView} from "./text_annotation"
 import {compute_angle, invert_angle, atan2} from "core/util/math"
 import type {CoordinateMapper} from "core/util/bbox"
 import {CoordinateUnits, AngleUnits, Direction} from "core/enums"
-import type {Size} from "core/layout"
 import {SideLayout} from "core/layout/side_panel"
 import type * as p from "core/properties"
-import type {GraphicsBox} from "core/graphics"
 import {assert} from "core/util/assert"
 import type {Pannable, PanEvent, KeyModifiers} from "core/ui_events"
 import {Signal} from "core/signaling"
 import {rotate_around} from "core/util/affine"
-import type {LRTB, XY, Corners} from "core/util/bbox"
-import {BBox} from "core/util/bbox"
-import {TextAnchor, Padding, BorderRadius} from "../common/kinds"
+import type {XY, SXY} from "core/util/bbox"
+import {TextAnchor} from "../common/kinds"
 import * as resolve from "../common/resolve"
-import {round_rect} from "../common/painting"
 import {Node} from "../coordinates/node"
 
 type HitTarget = "area"
-type SXY = {sx: number, sy: number}
 
 function xy<T>(x: T, y: T): XY<T> {
   return {x, y}
@@ -34,20 +29,6 @@ export class LabelView extends TextAnnotationView implements Pannable {
       this.layout = new SideLayout(panel, () => this.get_size(), false)
     else
       this.layout = undefined
-  }
-
-  protected override _get_size(): Size {
-    if (!this.displayed) {
-      return {width: 0, height: 0}
-    }
-
-    const graphics = this._text_view.graphics()
-    graphics.angle = this.angle
-    graphics.align = "auto"
-    graphics.visuals = this.visuals.text.values()
-
-    const {width, height} = graphics.size()
-    return {width, height}
   }
 
   get mappers(): XY<CoordinateMapper> {
@@ -78,14 +59,6 @@ export class LabelView extends TextAnnotationView implements Pannable {
     return resolve.text_anchor(this.model.anchor, align, baseline)
   }
 
-  get padding(): LRTB<number> {
-    return resolve.padding(this.model.padding)
-  }
-
-  get border_radius(): Corners<number> {
-    return resolve.border_radius(this.model.border_radius)
-  }
-
   get angle(): number {
     const {angle, angle_units, direction} = this.model
     return compute_angle(angle, angle_units, direction)
@@ -103,75 +76,6 @@ export class LabelView extends TextAnnotationView implements Pannable {
     const sy = compute("y", y, mappers.y) - y_offset // TODO this needs to be unified with the rest of bokehjs
 
     return {sx, sy}
-  }
-
-  protected _text_box: GraphicsBox
-  protected _rect: {
-    sx: number
-    sy: number
-    width: number
-    height: number
-    angle: number
-    anchor: XY<number>
-    padding: LRTB<number>
-    border_radius: Corners<number>
-  }
-
-  override compute_geometry(): void {
-    super.compute_geometry()
-
-    const text_box = this._text_view.graphics()
-    text_box.position = {sx: 0, sy: 0, x_anchor: "left", y_anchor: "top"}
-    text_box.angle = 0 // needs reset because text_box is self-referential
-    text_box.align = "auto"
-    text_box.visuals = this.visuals.text.values()
-
-    const size = text_box.size()
-    const {sx, sy} = this.origin
-    const {anchor, padding, border_radius, angle} = this
-
-    const width = size.width + padding.left + padding.right
-    const height = size.height + padding.top + padding.bottom
-
-    this._text_box = text_box
-    this._rect = {sx, sy, width, height, angle, anchor, padding, border_radius}
-  }
-
-  protected _render(): void {
-    this.compute_geometry() // TODO: remove this
-
-    const {ctx} = this.layer
-
-    const {sx, sy, width, height, angle, anchor, padding, border_radius} = this._rect
-    const label = this._text_box
-
-    const dx = anchor.x*width
-    const dy = anchor.y*height
-
-    ctx.translate(sx, sy)
-    ctx.rotate(angle)
-    ctx.translate(-dx, -dy)
-
-    const {background_fill, background_hatch, border_line, text} = this.visuals
-    if (background_fill.doit || background_hatch.doit || border_line.doit) {
-      ctx.beginPath()
-      const bbox = new BBox({x: 0, y: 0, width, height})
-      round_rect(ctx, bbox, border_radius)
-      background_fill.apply(ctx)
-      background_hatch.apply(ctx)
-      border_line.apply(ctx)
-    }
-
-    if (text.doit) {
-      const {left, top} = padding
-      ctx.translate(left, top)
-      label.paint(ctx)
-      ctx.translate(-left, -top)
-    }
-
-    ctx.translate(dx, dy)
-    ctx.rotate(-angle)
-    ctx.translate(-sx, -sy)
   }
 
   override interactive_hit(sx: number, sy: number): boolean {
@@ -267,8 +171,6 @@ export namespace Label {
     angle: p.Property<number>
     angle_units: p.Property<AngleUnits>
     direction: p.Property<Direction>
-    padding: p.Property<Padding>
-    border_radius: p.Property<BorderRadius>
     editable: p.Property<boolean>
   }
 
@@ -301,8 +203,6 @@ export class Label extends TextAnnotation {
       angle:       [ Angle, 0 ],
       angle_units: [ AngleUnits, "rad" ],
       direction:   [ Direction, "anticlock" ],
-      padding:     [ Padding, 0 ],
-      border_radius: [ BorderRadius, 0 ],
       editable:    [ Boolean, false ],
     }))
   }
